@@ -44,10 +44,10 @@
     "@mui/icons-material": "^5.3.0",
     "@mui/material": "^5.3.0",
     "dayjs": "^1.10.7",
-    "react": "^17.0.2",
+    "react": "^18.0.0-rc.0",
     "react-color": "^2.19.3",
     "react-day-picker": "^7.4.10",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0-rc.0",
     "react-router-dom": "^5.1.2",
     "uuid": "^8.3.2"
   }
